Fix bug causing final mission crash
Game crashes after the final mission cutscenes.

This is the usual screen listener holding onto object references bug
preventing them from being GC'd. We get no OnRemoved signal to the
listeners when transitioning to the final mission after the cutscene
movies, so we have no opportunity to clean up our held references as we
do for most missions.

The fix is to restructure the code so we no longer need to hold onto
cached references to the UISquadSelect_ListItems. Instead, create the
icons with well-known names and search the list items for sub-panels
with those names when deciding if we need to create new instances or
show/hide existing ones.