Add WotC detection
Adds two new config variables WOTC_(WEAPON|PCS)_X_POSITION for use with
WotC since the default position of the PCS icon from vanilla overlaps
with the soldier bond icon.

When creating the icons try to see if we're in WotC (by searching for
the named bond icon element in the soldier's list item) and use the
right config based on whether we find it or not.